Bug(answer route): create answer route and controller files [Finishes #159980780]
