+++ .gitignore
@@ -7,14 +7,8 @@ erl_crash.dump
 *.ez
 *.beam
 /config/*.secret.exs
-<<<<<<< HEAD
-/Joyce/
-node_modules/
-.nyc_output/
-=======
 node_modules/
 package-lock.json
 coverage
 .nyc_output/
 
->>>>>>> 3b574419a5d18abca0c1cbdf1f63e9f842b27444
