+++ server/controllers/QuestionsHandler.js
@@ -1,4 +1,4 @@
-import questions from '../sessionData/questionsObj';
+import questions from '../dummyData/questionsObj';
 
 class QuestionsHandler {
     static getAllQuestions (request, response) {
@@ -43,23 +43,6 @@ class QuestionsHandler {
           requestedQues
         });
     }
-  
-    static postAnswer(request, response) {
-      const { newAnswer, requestedQues } = request.body;
-      const id = requestedQues.answer.length + 1;
-      const submittedAnswer = {
-        id,
-        newAnswer
-      };
-      requestedQues.answer.push(submittedAnswer);
-      requestedQues.answer.reverse();
-      return response.status(201)
-        .json({
-          message: 'Thank you! Your answer was recorded',
-          requestedQues
-        });
-    }
-
     static deleteQuestion(request, response) {
       const { requestedQues } = request.body;
       questions.splice(requestedQues.id - 1, 1);
