+++ server/routes/index.js
@@ -1,0 +1,6 @@
+import answerRoute from './answerRoute';
+import questionRoute from './questionsRoute';
+
+export {
+    answerRoute, questionsRoute
+}
