+++ server/routes/questionsRoute.js
@@ -10,7 +10,6 @@ const router = express.Router();
 router.get('/questions', getAllQuestions);
 router.get('/questions/:questionId', getQuestionValidator, getQuestion);
 router.post('/questions', postQuestionValidator, postQuestion);
-router.post('/questions/:questionId/answers', getQuestionValidator, postAnswerValidator, postAnswer);
 router.delete('/questions/:questionId', getQuestionValidator, deleteQuestion);
 
 router.get('/', (request, response) => {
