+++ server/tests/questionsTest.js
@@ -1,85 +1,14 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import app from '../../app';
-import questions from '../sessionData/questionsObj';
+import questions from '../dummyData/questionsObj';
+import {
+    correctQuestion, undefinedTitle, undefinedContent, undefinedTag, emptyTitle, emptyContent, emptyTag, titleLength, contentLength, tagLenght, correctAnswer, undefinedAnswer, emptyAnswer, answerLength
+} from '../dummyData/mockQuestions';
 
 const { expect } = chai;
 
 chai.use(chaiHttp);
-
-const correctQuestion = {
-    title: 'How to merge to git',
-    content: 'How do I merge two different git branches to master?',
-    tag: 'git'
-};
-
-const undefinedTitle = {
-    content: 'How do I merge two different git branches to master?',
-    tag: 'git'
-};
-
-const undefinedContent = {
-    title: 'How to merge to git',
-    tag: 'git'
-};
-
-const undefinedTag = {
-    title: 'How to merge to git',
-    content: 'How do I merge two different git branches to master?'
-};
-
-const emptyTitle = {
-    title: '',
-    content: 'How do I merge two different git branches to master?',
-    tag: 'git',
-};
-
-const emptyContent = {
-    title: 'How to merge to git',
-    content: '',
-    tag: 'git',
-};
-
-const emptyTag = {
-    title: 'How to merge to git',
-    content: 'How do I merge two different git branches to master?',
-    tag: '',
-};
-
-const titleLength = {
-    title: 'ow',
-    content: 'How do I merge two different git branches to master?',
-    tag: 'git',
-};
-
-const contentLength = {
-    title: 'How to merge to git',
-    content: 'How',
-    tag: 'git',
-};
-
-const tagLenght = {
-    title: 'How to merge to git',
-    content: 'How do I merge two different git branches to master?',
-    tag: 'it',
-};
-
-const correctAnswer = {
-    newAnswer: 'Check link for answer'
-};
-
-const undefinedAnswer = {
-
-};
-
-const emptyAnswer = {
-    newAnswer: ''
-};
-
-const answerLength = {
-    newAnswer: 'ik'
-};
-
 
 describe("StackOverflow-lite", () => {
